requirements.txt
requests~=2.28.2
primer3-py~=1.0.0